gce_image_publish: move publish type into its own package (#695)
Move publish and image struct into a separate package, so we can programatically fill up publish/image and marshal into json and then feed into the publish tool
