+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/GoogleCloudPlatform/compute-image-tools
+
+go 1.12
+
+require (
+	cloud.google.com/go v0.36.0
+	github.com/kylelemons/godebug v0.0.0-20170820004349-d65d576e9348
+	golang.org/x/oauth2 v0.0.0-20190226205417-e64efc72b421
+	google.golang.org/api v0.1.0
+)
